% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_geocoordinate.R
\name{city_geocoordinate}
\alias{city_geocoordinate}
\title{city_geocoordinate}
\usage{
city_geocoordinate(name)
}
\arguments{
\item{name}{Name of a city}
}
\value{
list: latitude, longitude, (OSM) display name, raw.list returned by API
}
\description{
Retrieve geocoordinates for one city(/town/...) name from OSM Nominatim API. WIP
}
\details{
Tweaked version of gisfin/geocode.R authored by Juuso Parkkinen.
}
\examples{
\dontrun{TODO}
}
\author{
Aaro Salosensaari \email{aaro.salosensaari@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

