% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_geocoordinates_all.R
\name{city_geocoordinates_all}
\alias{city_geocoordinates_all}
\title{city_geocoordinates_all}
\usage{
city_geocoordinates_all(names, wait = 1)
}
\arguments{
\item{names}{A list of city names}

\item{wait}{Courtesy wait (seconds) between requests to Nominatim API.}
}
\value{
DF with two columns (lat, long)
}
\description{
Retrieve geocoordinates for a list of cities from OSM Nominatim API. WIP
}
\details{
TODO
}
\examples{
TODO
}
\author{
Aaro Salosensaari \email{aaro.salosensaari@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

