% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_synonymes.R
\name{city_synonymes}
\alias{city_synonymes}
\title{city_synonymes}
\usage{
city_synonymes(unique_cities)
}
\arguments{
\item{unique_cities}{List of unique city names in the data}
}
\value{
List of city name variants
}
\description{
Use Levenshtein distance iteratively to find plausible misspellings for specific cities
}
\examples{
\dontrun{all <- city_synonymes(c("Turku","Turkue", "Tampere", "Helsinki"}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

