% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_publisher_fennica.R
\name{clean_publisher_fennica}
\alias{clean_publisher_fennica}
\title{Clean publisher}
\usage{
clean_publisher_fennica(x, languages = c("english"))
}
\arguments{
\item{x}{Vector of publisher names}

\item{languages}{A vector of languages which are used in detecting relation keywords}
}
\value{
Vector of publisher names
}
\description{
Pre-cleans publisher field
}
\examples{
# clean_publisher(x, languages=c("finnish", "swedish", "latin"))
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

