% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/city.R
\name{collect_misspellings}
\alias{collect_misspellings}
\title{collect_misspellings}
\usage{
collect_misspellings(city, variations, all)
}
\arguments{
\item{city}{Correct name (prechosen form)}

\item{variations}{List of usual forms (possibly including usual misspellings)}

\item{all}{List of all non-assigned city names in the data}
}
\value{
List of all non-assigned city names in the data
}
\description{
Use Levenshtein distance iteratively to find plausible misspellings
}
\examples{
\dontrun{all <- collect_misspellings("Hämeenlinna",c("Hämeenlinna","Tavastehus"),c("Hämeenlinnae","Tavasteehuus"))}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

