% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{correct_misspellings}
\alias{correct_misspellings}
\title{correct_misspellings}
\usage{
correct_misspellings(v, city)
}
\arguments{
\item{v}{Vector of city names}

\item{city}{Correct name of a city}
}
\value{
Vector of city names
}
\description{
Correct misspelled city names based on mapping constructed earlier
}
\examples{
\dontrun{v <- correct_misspellings(v,"Hämeenlinna")}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

