% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{counts_by_language}
\alias{counts_by_language}
\title{counts_by_language}
\usage{
counts_by_language(df, str, file)
}
\arguments{
\item{df}{Main dataframe}

\item{str}{Name of town}

\item{file}{Filename}
}
\description{
Plot publication year counts by language
}
\examples{
\dontrun{counts_by_language(df,"finnish","lang_finnish")}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

