% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{counts_by_town}
\alias{counts_by_town}
\title{counts_by_town}
\usage{
counts_by_town(df, str, file)
}
\arguments{
\item{df}{Main dataframe}

\item{str}{Name of town}

\item{file}{Filename}
}
\description{
Plot publication year counts by town
}
\examples{
\dontrun{counts_by_town(df,"Hämeenlinna","Hameenlinna")}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

