% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/country.R
\name{deduce_country}
\alias{deduce_country}
\title{deduce_country}
\usage{
deduce_country(df)
}
\arguments{
\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Add country field to data frame based on city field
}
\examples{
\dontrun{df <- deduce_country(df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

