% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editing.R
\name{embellish}
\alias{embellish}
\title{embellish}
\usage{
embellish(v)
}
\arguments{
\item{v}{Vector of strings}
}
\value{
Vector of strings
}
\description{
Remove preceding and following characters from strings
}
\examples{
v <- embellish(c("  Bernard Mandeville         "," John Locke    "))
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

