% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{fix_birth}
\alias{fix_birth}
\title{fix_birth}
\usage{
fix_birth(df.orig, df)
}
\arguments{
\item{df.orig}{Dataframe with original fields}

\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Pick and polish birth year
}
\examples{
\dontrun{df <- fix_birth(df.orig, df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

