% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{fix_death}
\alias{fix_death}
\title{fix_death}
\usage{
fix_death(df.orig, df)
}
\arguments{
\item{df.orig}{Dataframe with original fields}

\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Pick and polish year of death
}
\examples{
\dontrun{df <- fix_death(df.orig, df)}
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

