% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notes.R
\name{fix_note_diss}
\alias{fix_note_diss}
\title{fix_note_diss}
\usage{
fix_note_diss(df.orig, df)
}
\arguments{
\item{df.orig}{Dataframe with original fields}

\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Establish two binary variables: dissertation and synodal dissertation
}
\examples{
\dontrun{df <- fix_note_diss(df.orig, df)}
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

