% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.R
\name{fix_pubfrom}
\alias{fix_pubfrom}
\title{fix_pubfrom}
\usage{
fix_pubfrom(df.orig, df)
}
\arguments{
\item{df.orig}{Dataframe with original fields}

\item{df}{Dataframe with modified fields}
}
\value{
Dataframe with modified fields
}
\description{
Pick and polish first year of publication (newspapers and magazines)
}
\examples{
\dontrun{df <- fix_pubfrom(df.orig, df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

