% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time.R
\name{fix_pubtill}
\alias{fix_pubtill}
\title{fix_pubtill}
\usage{
fix_pubtill(df.orig, df)
}
\arguments{
\item{df.orig}{Dataframe with original fields}

\item{df}{Dataframe with modified fields}
}
\value{
Dataframe with modified fields
}
\description{
Pick and polish last year of publication (newspapers and magazines)
}
\examples{
\dontrun{df <- fix_pubtill(df.orig, df)}
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

