% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/time.R
\name{fix_pubwhen}
\alias{fix_pubwhen}
\title{fix_pubwhen}
\usage{
fix_pubwhen(df.orig, df)
}
\arguments{
\item{df.orig}{Dataframe with original fields}

\item{df}{Dataframe with modified fields}
}
\value{
Dataframe with modified fields
}
\description{
Pick and polish year of publication (books)
}
\examples{
\dontrun{df <- fix_pubwhen(df.orig, df)}
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

