% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notes.R
\name{fix_university}
\alias{fix_university}
\title{fix_university}
\usage{
fix_university(df.orig, df)
}
\arguments{
\item{df.orig}{Dataframe with original fields}

\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Polish university names
}
\examples{
\dontrun{df <- fix_university(df.orig, df)}
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

