% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publishers_Finto.R
\name{get_publishers_Finto}
\alias{get_publishers_Finto}
\title{Unify Publishers with Finto}
\usage{
get_publishers_Finto(
  cheat_list,
  Finto_comp,
  all_names,
  unknown_town,
  publication_year
)
}
\arguments{
\item{cheat_list}{cheat_list}

\item{Finto_comp}{Vector of alternative names from Finto}

\item{all_names}{Data frame containing all the name related fields}

\item{unknown_town}{Vector of towns from raw data}

\item{publication_year}{Data frame of published_in, published_from and published_till}
}
\value{
Data frame with alt, pref and match_methods
}
\description{
Use Finto database to unify publisher names.
}
\references{
See citation("fennica")
}
\author{
Hege Roivainen \email{leo.lahti@iki.fi}
}
\keyword{utilities}
