% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_corporate_Finto.R
\name{harmonize_corporate_Finto}
\alias{harmonize_corporate_Finto}
\title{Unify Publisher names to process with Finto}
\usage{
harmonize_corporate_Finto(x)
}
\arguments{
\item{x}{Vector of publishers names}
}
\value{
Data frame with orig, name, town, year_from and year_till
}
\description{
Use Finto database to unify publisher names.
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
