% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publisher_fennica.R
\name{harmonize_publisher_fennica}
\alias{harmonize_publisher_fennica}
\title{Harmonize Publisher Fennica}
\usage{
harmonize_publisher_fennica(df.orig, cheat_list, languages = c("english"),
  useShortcuts = FALSE)
}
\arguments{
\item{df.orig}{Raw data from csv}

\item{cheat_list}{Data frame of alt, pref: alternative spellings and their counterpart the preferred spelling}

\item{languages}{A vector of languages which are used in detecting relation keywords}

\item{useShortcuts}{If TRUE, loads RDS instead of doing everything from scratch}
}
\value{
Data frame with alt, pref, match_methods
}
\description{
Harmonizes publisher names and enriches them with Finto information.
}
\examples{
# harmonize_publisher_fennica(df.orig, cheat_list, languages=c("finnish", "swedish", "latin"))
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
