% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publisher_fennica_part2.R
\name{harmonize_publisher_fennica_part2}
\alias{harmonize_publisher_fennica_part2}
\title{Harmonize Publisher Fennica}
\usage{
harmonize_publisher_fennica_part2(df.orig)
}
\arguments{
\item{df.orig}{Data frame with raw data, assuming the place, year fields are already polished.}
}
\value{
A vector with polished entries.
}
\description{
Main handler for publisher fields for Fennica.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

