% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/editing.R
\name{paste_without_NA}
\alias{paste_without_NA}
\title{paste_without_NA}
\usage{
paste_without_NA(x1 = rep(NA, data_size), x2 = rep(NA, data_size))
}
\arguments{
\item{x1}{Vector of strings}

\item{x2}{Vector of strings}
}
\value{
Vector of strings
}
\description{
Variant of two-variable paste for situations where NA is possible
}
\examples{
t <- paste_without_NA()
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

