% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_entry.R
\name{polish_entry}
\alias{polish_entry}
\title{Polish Entry}
\usage{
polish_entry(x)
}
\arguments{
\item{x}{Vector of entries}
}
\value{
Vector of entries polished
}
\description{
Simple function for minimal field harmonization.
}
\details{
Remove ending commas, periods, spaces and parentheses, 
	    starting prepositions etc.
}
\examples{
\dontrun{x2 <- polish_title(x)}
}
\references{
See citation("fennica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
