% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_languages.R
\name{polish_languages}
\alias{polish_languages}
\title{Polish Languages}
\usage{
polish_languages(x)
}
\arguments{
\item{x}{language field (a vector)}
}
\value{
data.frame with separate fields for different languages
}
\description{
Construct binary matrix of languages for each entry
}
\examples{
\dontrun{df <- mark_languages(c("fin;lat","eng"))}
}
\references{
See citation("fennica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
