% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publisher_fennica.R
\name{polish_publisher_fennica}
\alias{polish_publisher_fennica}
\title{Polish Publishers for Fennica}
\usage{
polish_publisher_fennica(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
Data frame with possible publisher form and the preferred output
}
\description{
Fennica-specific cleanup for the publisher field.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}

