% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/city.R
\name{preprocess_placenames}
\alias{preprocess_placenames}
\title{preprocess_placenames}
\usage{
preprocess_placenames(df.orig, df)
}
\arguments{
\item{df.orig}{Dataframe with original fields}

\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Polish placenames (preliminary step)
}
\examples{
\dontrun{df <- preprocess_placenames(df.orig, df)}
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

