% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{read_fennica_data}
\alias{read_fennica_data}
\title{read_fennica_data}
\usage{
read_fennica_data(file)
}
\arguments{
\item{file}{Source Fennica data file}
}
\value{
Dataframe with unpolished fields
}
\description{
Read preparsed Fennica data
}
\examples{
\dontrun{df.raw <- read_fennica_data(file)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

