% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/city.R
\name{sanitize_name}
\alias{sanitize_name}
\title{sanitize_name}
\usage{
sanitize_name(name)
}
\arguments{
\item{name}{String or vector of strings}
}
\value{
String or vector of strings
}
\description{
Remove Nordic characters etc. from strings to use in filenames
}
\examples{
name <- sanitize_name("Hämeenlinna")
}
\author{
Niko Ilomäki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

