% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_synonymes.R
\name{city_synonymes}
\alias{city_synonymes}
\title{city_synonymes}
\usage{
city_synonymes(x, synonymes, verbose = TRUE)
}
\arguments{
\item{x}{List of city names in the data}

\item{synonymes}{Table of known city name variants with the columns "name" and "synonyme"}

\item{verbose}{verbose}
}
\value{
List of city name variants
}
\description{
Use Levenshtein distance iteratively to find plausible misspellings for specific cities
}
\examples{
\dontrun{all <- city_synonymes(c("Turku","Turkue", "Tampere", "Helsinki"}
}
\author{
Leo Lahti and Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

