% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_total.R
\name{counts_total}
\alias{counts_total}
\title{counts_total}
\usage{
counts_total(df)
}
\arguments{
\item{df}{Main dataframe}
}
\description{
Plot publication year counts
}
\examples{
\dontrun{counts_total(df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

