% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publisher_fennica.R
\name{harmonize_publisher_fennica}
\alias{harmonize_publisher_fennica}
\title{Harmonize Publisher Fennica}
\usage{
harmonize_publisher_fennica(df, cheat_list, languages = c("english"))
}
\arguments{
\item{df}{data.frame}

\item{cheat_list}{Data frame of alt, pref: alternative spellings and their counterpart the preferred spelling}

\item{languages}{A vector of languages which are used in detecting relation keywords}
}
\value{
Polished vector.
}
\description{
Harmonizes publisher names and enriches them with Finto information.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

