% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publisher_fennica.R
\name{polish_publisher_fennica}
\alias{polish_publisher_fennica}
\title{Polish Publisher Main}
\usage{
polish_publisher_fennica(df.preprocessed, useShortcuts = FALSE)
}
\arguments{
\item{df.preprocessed}{Data frame with raw data}

\item{useShortcuts}{If TRUE, loads RDS instead of doing everything from scratch}
}
\value{
Data frame with orig, mod
}
\description{
Main handler for publisher fields.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

