% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publisher_fennica.R
\name{polish_publisher_fennica}
\alias{polish_publisher_fennica}
\title{Harmonize Publisher Main}
\usage{
polish_publisher_fennica(df)
}
\arguments{
\item{df}{Data frame with raw data}
}
\value{
Data frame with orig, mod
}
\description{
Main handler for publisher fields.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

