% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_name.R
\name{sanitize_name}
\alias{sanitize_name}
\title{sanitize_name}
\usage{
sanitize_name(x)
}
\arguments{
\item{x}{String or vector of strings}
}
\value{
String or vector of strings
}
\description{
Remove Nordic characters etc. from strings to use in filenames
}
\examples{
name <- sanitize_name("Turku")
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("fennica")
}
\keyword{utilities}

